refac: Move price getter function into utils
